requirements.txt
aiohttp==3.8.3
aiolinebot==0.4.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.89.1
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
httptools==0.5.0
idna==3.4
line-bot-sdk==2.4.1
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
pydantic==1.10.4
PyMySQL==1.0.2
python-decouple==3.7
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.1
shortuuid==1.0.11
sniffio==1.3.0
SQLAlchemy==2.0.2
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
utils==1.0.1
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
